requirements.txt
websocket-client>=1.2.1
requests>=2.28.1